requirements.txt
alabaster==0.7.8
amqp==2.5.1
asn1crypto==0.24.0
audioread==2.1.8
auto-py-torrent==1.0.8
Babel==2.4.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.6.1.0
celery==4.3.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
coloredlogs==10.0
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-celery-beat==1.5.0
django-ordered-model==3.3.0
django-timezone-field==3.0
docutils==0.14
easygui==0.96
entrypoints==0.2.3.post1
humanfriendly==4.18
idna==2.6
image==1.5.27
imagesize==1.0.0
importlib-metadata==0.20
Jinja2==2.10
joblib==0.13.2
keyring==12.2.1
keyrings.alt==3.0
kiwisolver==1.1.0
kombu==4.6.4
librosa==0.7.0
llvmlite==0.29.0
lxml==4.3.3
MarkupSafe==1.0
matplotlib==3.1.1
more-itertools==7.2.0
nltk==3.4.5
numba==0.45.1
numpy==1.17.0
onboard==1.4.1
opencv-python==4.1.0.25
packaging==17.1
paramiko==2.6.0
persepolis==3.0.1
pexpect==4.2.1
Pillow==5.1.0
pocketsphinx==0.1.15
psutil==5.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygame==1.9.5
Pygments==2.2.0
pygobject==3.28.2
PyNaCl==1.3.0
pyparsing==2.4.1.1
pysmbc==1.0.15.6
python-apt==1.6.0
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
pyttsx3==2.71
pytz==2019.1
pyxdg==0.25
reportlab==3.4.0
requests==2.18.4
resampy==0.2.2
roman==2.0.0
scikit-learn==0.21.3
scipy==1.3.1
SecretStorage==2.3.1
selenium==3.141.0
setproctitle==1.1.10
six==1.11.0
Slowloris==0.1.5
SoundFile==0.10.2
soupsieve==1.9.1
speake3==0.3
speech==0.5.2
SpeechRecognition==3.8.1
Sphinx==1.7.4
sqlparse==0.3.0
tabulate==0.8.3
tk==0.1.0
unattended-upgrades==0.1
urllib3==1.22
vboxapi==1.0
vine==1.3.0
wget==3.2
whitenoise==4.1.3
zipp==0.6.0
